adding redux functions and axios calls, separating out redux actions into separate files
