+++ recruiter-front-end/src/App.js
@@ -12,7 +12,7 @@ import Education from './views/NewRulesPage/Education';
 import Skills from './views/NewRulesPage/Skills';
 import Experience from './views/NewRulesPage/Experience';
 
-import Contacts from './components/Contacts/NewContactForm';
+import NewContactForm from './components/Contacts/NewContactForm';
 import Mailer from './components/Mailer';
 import fire from './config/fire';
 import { Menu, Button, Container } from 'semantic-ui-react';
