+++ recruiter-front-end/src/actions/emailActions.js
@@ -1,0 +1,16 @@
+import axios from 'axios';
+
+// TOKEN
+
+const token = sessionStorage.getItem('token');
+const tokenHeader = {
+  headers: {
+    token: token,
+  },
+};
+
+// GET EMAIL HISTORY
+
+// ADD EMAIL HISTORY
+
+// SEND EMAIL
