+++ recruiter-front-end/src/components/Contacts/Contacts.js
@@ -30,7 +30,7 @@ class Contacts extends React.Component {
     })
       .then(res => this.setState({ contacts: res.data }))
       .catch(error => console.log(error));
-
+  };
   render() {
     this.state.contacts.length < 1 ? (
       <Container fluid>
